인증체크 완료

+++ src/main/java/site/metacoding/white/web/BoardApiController.java
@@ -26,11 +26,13 @@ public class BoardApiController {
     private final HttpSession session;
 
     @PostMapping("/board")
-    public ResponseDto<Object> save(@RequestBody BoardSaveReqDto boardSaveReqDto) { // json으로 받기 위해서 @RequestBody
+    public ResponseDto<?> save(@RequestBody BoardSaveReqDto boardSaveReqDto) {
         SessionUser sessionUser = (SessionUser) session.getAttribute("sessionUser");
-        // insert into board(title,content,user_id) values(?,?,?)
+        if (sessionUser == null) {
+            throw new RuntimeException("로그인이 필요합니다.");
+        }
         boardSaveReqDto.setSessionUser(sessionUser);
-        BoardSaveRespDto boardSaveRespDto = boardService.save(boardSaveReqDto); // 서비스에는 단 하나의 객체만 전달한다
+        BoardSaveRespDto boardSaveRespDto = boardService.save(boardSaveReqDto); // 서비스에는 단 하나의 객체만 전달한다.
         return new ResponseDto<>(1, "성공", boardSaveRespDto);
     }
 
@@ -47,12 +49,20 @@ public class BoardApiController {
 
     @PutMapping("/board/{id}")
     public ResponseDto<?> update(@PathVariable Long id, @RequestBody BoardUpdateReqDto boardUpdateReqDto) {
+        SessionUser sessionUser = (SessionUser) session.getAttribute("sessionUser");
+        if (sessionUser == null) {
+            throw new RuntimeException("로그인이 필요합니다.");
+        }
         boardUpdateReqDto.setId(id);
         return new ResponseDto<>(1, "성공", boardService.update(boardUpdateReqDto));
     }
 
     @DeleteMapping("/board/{id}")
     public ResponseDto<?> deleteById(@PathVariable Long id) {
+        SessionUser sessionUser = (SessionUser) session.getAttribute("sessionUser");
+        if (sessionUser == null) {
+            throw new RuntimeException("로그인이 필요합니다.");
+        }
         boardService.deleteById(id);
         return new ResponseDto<>(1, "성공", null);
     }
